requirements.txt
# Core audio processing
vosk==0.3.44
openai-whisper==20231117
sounddevice==0.4.6
torch==2.2.2
numpy==1.26.4
tqdm==4.66.2

# GUI
customtkinter==5.2.2
Pillow==10.2.0
opencv-python-headless==4.9.0.80

# Local AI & Control
python-rtmidi==1.5.8

# Utilities
loguru==0.7.2

# NDI Video Output (Optional, requires manual SDK installation)
# pysimplendi==0.1.3

certifi==2024.2.2
charset-normalizer==3.3.2
requests==2.31.0
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
whisper==1.1.10
python-dotenv==0.21.1
python-dotenv
python-dotenv
